virtio: Replace size return with usedLength in queue_dequeue.
The size was in fact the count of physical entries that were used. That
number must necessarily be the same as the number given when adding to
the queue, so that number isn't really interesting. Consequently none
of the users of that API made use of it.

Return the used length instead, which is the way virtio signals how much
valid data resides in the dequeued buffer. This is for example important
to know the frame length of incoming packets in virtio_net.

Change-Id: I00af89d391c7df4dd8b3b39c22a4505e593af938
Reviewed-on: https://review.haiku-os.org/721
Reviewed-by: waddlesplash <waddlesplash@gmail.com>