HDA: VMware's emulated controllers have HDA_QUIRK_NO_CORBRP_RESET_ACK.
On my VMware Player installation, the device ID is "1977", but presumably
VMware will have only one HDA codebase, and so we just enable the fix on
all devices.

This fixes initialization of the HDA driver on VMware, and means that
it is now possible to get sound output on VMware without using the
OpenSound driver package.