runtime_loader: Make the heap thread-safe.
Previously we didn't need to care about threading in here, as it
was used only up to the point where the application's _start was called
and then libroot's heap would take over. But with the introduction of TLS
support by pdziepak in 2014, we also handle TLS bookkeeping inside
runtime_loader, and so this heap needs to be thread-safe.

Properly fixes the JVM crashes korli was attempting to fix in hrev52658,
and fixes #13154, #14129, #14304, #14342 for real.

Change-Id: Ia3398ab47c4f7f290aef41982775bfea05230589
Reviewed-on: https://review.haiku-os.org/777
Reviewed-by: waddlesplash <waddlesplash@gmail.com>